+++ kotlin-hocon-mapper/src/test/kotlin/com/github/uharaqo/hocon/mapper/UnitConversionModels.kt
@@ -12,7 +12,7 @@ import kotlinx.serialization.UseSerializers
 import java.time.Duration
 import java.time.Period
 
-interface AdditionalModels {
+interface UnitConversionModels {
 
     @Serializable
     data class Periods(
